backend/requirements.txt
streamlit
pymupdf
docling-parse
# pdfplumber
"magic-pdf[full]"
"unstructured[pdf]"
pypdf
sycamore-ai
loguru